feat: headless webview for darwin platform
